move printPartitions to application
